+++ src/app/[locale]/admin/events/calendar/_components/useAdminCreateEventForm.tsx
@@ -29,7 +29,6 @@ import type {
   PrepareResponse,
   PreSignCheckKey,
   ProposerWhitelistCheckState,
-  RecurringOccurrencePreview,
   ResolutionType,
   SignatureExecutionTx,
   SignerOption,
@@ -66,18 +65,10 @@ import {
 } from '@/lib/admin-sports-create'
 import { formatDateTimeLocalValue, normalizeDateTimeLocalValue } from '@/lib/datetime-local'
 import {
-  addRecurrenceInterval,
   appendEventCreationSlugSuffix,
-  applyEventCreationTemplate,
-  buildDefaultDeployAt,
-  buildEventCreationTimestampSeed,
   buildEventCreationWalletTail,
-  buildImmediateDeployAt,
-  buildScheduledRecurringDeployAt,
-  hasEventCreationDateTemplateVariable,
   normalizeEventCreationAssetPayload,
   slugifyEventCreationValue as slugify,
-  slugifyEventCreationTemplate as slugifyTemplate,
 } from '@/lib/event-creation'
 import {
   isProposerWhitelistStatusResponse,
@@ -136,7 +127,6 @@ import {
   fetchAdminApiWithTimeout,
   getAiIssueKey,
   getChainLabel,
-  hasRecurringDeploymentHistory,
   isAiValidationResponse,
   isAlreadyInitializedError,
   isBigIntSerializationError,
@@ -160,6 +150,8 @@ import { useAdminWalletActions } from './useAdminWalletActions'
 import { useAiRules } from './useAiRules'
 import { useAllowedCreatorWallets } from './useAllowedCreatorWallets'
 import { useEventAssets } from './useEventAssets'
+import { useEventPreview } from './useEventPreview'
+import { useRecurringEventPreview } from './useRecurringEventPreview'
 import { useSignatureCountdown } from './useSignatureCountdown'
 import { useSportsMarketRows } from './useSportsMarketRows'
 import { useSportsMatchSearch } from './useSportsMatchSearch'
@@ -687,309 +679,61 @@ export function useAdminCreateEventForm({
     () => removeGeneratedCategoryItems(form.categories, sportsGeneratedCategorySlugs),
     [form.categories, sportsGeneratedCategorySlugs],
   )
-  const scheduleDateValue = useMemo(
-    () => normalizeDateTimeLocalValue(form.endDateIso),
-    [form.endDateIso],
-  )
-  const scheduleOccurrenceDate = useMemo(() => {
-    if (!scheduleDateValue) {
-      return null
-    }
-
-    const parsed = new Date(scheduleDateValue)
-    return Number.isNaN(parsed.getTime()) ? null : parsed
-  }, [scheduleDateValue])
-  const recurrenceIntervalNumber = useMemo(
-    () => Math.max(1, Number.parseInt(recurrenceInterval || '1', 10) || 1),
-    [recurrenceInterval],
-  )
-  const hasRecurringDeployHistory = useMemo(
-    () => creationMode === 'recurring' && hasRecurringDeploymentHistory(initialDraftRecord),
-    [creationMode, initialDraftRecord],
-  )
-  const automaticDeployAtIso = useMemo(() => {
-    if (!scheduleOccurrenceDate) {
-      return null
-    }
-
-    if (creationMode !== 'recurring') {
-      return buildDefaultDeployAt(scheduleOccurrenceDate)?.toISOString() ?? null
-    }
-
-    if (!hasRecurringDeployHistory) {
-      return buildImmediateDeployAt(clientNowMs)?.toISOString() ?? null
-    }
-
-    return buildScheduledRecurringDeployAt(
-      scheduleOccurrenceDate,
-      recurrenceUnit || null,
-      recurrenceIntervalNumber,
-    )?.toISOString() ?? null
-  }, [clientNowMs, creationMode, hasRecurringDeployHistory, recurrenceIntervalNumber, recurrenceUnit, scheduleOccurrenceDate])
-  const automaticDeployAtDate = useMemo(() => {
-    if (!automaticDeployAtIso) {
-      return null
-    }
-
-    const parsed = new Date(automaticDeployAtIso)
-    return Number.isNaN(parsed.getTime()) ? null : parsed
-  }, [automaticDeployAtIso])
-  const nextRecurringResolutionDate = useMemo(() => {
-    if (creationMode !== 'recurring' || !scheduleOccurrenceDate || !recurrenceUnit) {
-      return null
-    }
-
-    return addRecurrenceInterval(scheduleOccurrenceDate, recurrenceUnit, recurrenceIntervalNumber)
-  }, [creationMode, recurrenceIntervalNumber, recurrenceUnit, scheduleOccurrenceDate])
-  const nextRecurringDeployDate = useMemo(() => {
-    if (!nextRecurringResolutionDate || !recurrenceUnit) {
-      return null
-    }
-
-    return buildScheduledRecurringDeployAt(nextRecurringResolutionDate, recurrenceUnit, recurrenceIntervalNumber)
-  }, [nextRecurringResolutionDate, recurrenceIntervalNumber, recurrenceUnit])
-  const recurringResolvedTitle = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return ''
-    }
-
-    const baseTemplate = titleTemplate.trim()
-    if (!baseTemplate) {
-      return ''
-    }
-
-    if (!scheduleOccurrenceDate) {
-      return baseTemplate
-    }
-
-    return applyEventCreationTemplate(baseTemplate, scheduleOccurrenceDate, baseTemplate).trim() || baseTemplate
-  }, [creationMode, scheduleOccurrenceDate, titleTemplate])
-  const derivedRecurringSlugTemplate = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return ''
-    }
-
-    return slugifyTemplate(titleTemplate)
-  }, [creationMode, titleTemplate])
-  const recurringSlugSuffix = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return ''
-    }
-
-    const timestampSeed = scheduleOccurrenceDate
-      ? buildEventCreationTimestampSeed(scheduleOccurrenceDate)
-      : slugSeed
-
-    return `${timestampSeed}${creatorSlugTail}`
-  }, [creationMode, creatorSlugTail, scheduleOccurrenceDate, slugSeed])
-  const effectiveRecurringSlugTemplate = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return ''
-    }
-
-    return slugTemplate.trim() || derivedRecurringSlugTemplate
-  }, [creationMode, derivedRecurringSlugTemplate, slugTemplate])
-  const recurringResolvedSlug = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return ''
-    }
-
-    const baseTemplate = effectiveRecurringSlugTemplate
-      || slugify(recurringResolvedTitle)
-    if (!baseTemplate) {
-      return ''
-    }
-
-    const rawSlug = scheduleOccurrenceDate
-      ? applyEventCreationTemplate(baseTemplate, scheduleOccurrenceDate, baseTemplate)
-      : baseTemplate
-
-    if (!scheduleOccurrenceDate) {
-      return appendEventCreationSlugSuffix(baseTemplate, recurringSlugSuffix)
-    }
-
-    return appendEventCreationSlugSuffix(slugify(rawSlug || baseTemplate), recurringSlugSuffix)
-  }, [creationMode, effectiveRecurringSlugTemplate, recurringResolvedTitle, recurringSlugSuffix, scheduleOccurrenceDate])
-  const recurringResolvedRules = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return ''
-    }
-
-    const baseTemplate = form.resolutionRules.trim()
-    if (!baseTemplate) {
-      return ''
-    }
-
-    if (!scheduleOccurrenceDate) {
-      return baseTemplate
-    }
-
-    return applyEventCreationTemplate(baseTemplate, scheduleOccurrenceDate, baseTemplate).trim() || baseTemplate
-  }, [creationMode, form.resolutionRules, scheduleOccurrenceDate])
-  const effectiveResolutionRules = useMemo(
-    () => creationMode === 'recurring'
-      ? (recurringResolvedRules || form.resolutionRules.trim())
-      : form.resolutionRules.trim(),
-    [creationMode, form.resolutionRules, recurringResolvedRules],
-  )
-  const buildRecurringOccurrencePreview = useCallback((date: Date | null): RecurringOccurrencePreview | null => {
-    if (creationMode !== 'recurring' || !date) {
-      return null
-    }
-
-    const rawTitleTemplate = titleTemplate.trim()
-    const resolvedTitle = applyEventCreationTemplate(rawTitleTemplate, date, rawTitleTemplate).trim()
-      || rawTitleTemplate
-      || form.title.trim()
-
-    const rawSlugTemplate = (effectiveRecurringSlugTemplate || slugify(resolvedTitle)).trim()
-    const resolvedBaseSlug = slugify(applyEventCreationTemplate(rawSlugTemplate, date, rawSlugTemplate) || rawSlugTemplate)
-    const suffix = `${buildEventCreationTimestampSeed(date)}${creatorSlugTail}`
-    const resolvedSlug = appendEventCreationSlugSuffix(resolvedBaseSlug, suffix)
-    const rawRulesTemplate = form.resolutionRules.trim()
-    const resolvedRules = applyEventCreationTemplate(rawRulesTemplate, date, rawRulesTemplate).trim() || rawRulesTemplate
-
-    return {
-      endDateIso: date.toISOString(),
-      title: resolvedTitle,
-      slug: resolvedSlug,
-      resolutionRules: resolvedRules,
-    }
-  }, [creationMode, creatorSlugTail, effectiveRecurringSlugTemplate, form.resolutionRules, form.title, titleTemplate])
-  const recurringOccurrencePreviews = useMemo(
-    () => creationMode === 'recurring'
-      ? [buildRecurringOccurrencePreview(scheduleOccurrenceDate), buildRecurringOccurrencePreview(nextRecurringResolutionDate)].filter(Boolean) as RecurringOccurrencePreview[]
-      : [],
-    [buildRecurringOccurrencePreview, creationMode, nextRecurringResolutionDate, scheduleOccurrenceDate],
-  )
-  const recurringPreviewErrors = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return [] as string[]
-    }
-
-    const errors: string[] = []
-    const [currentPreview, nextPreview] = recurringOccurrencePreviews
-
-    if (scheduleOccurrenceDate && !currentPreview?.slug) {
-      errors.push('Recurring slug preview is invalid.')
-    }
-
-    if (scheduleOccurrenceDate && !currentPreview?.title) {
-      errors.push('Recurring title preview is invalid.')
-    }
-
-    if (scheduleOccurrenceDate && !currentPreview?.resolutionRules) {
-      errors.push('Recurring resolution rules preview is invalid.')
-    }
-
-    if (currentPreview && nextPreview && currentPreview.slug === nextPreview.slug) {
-      errors.push('Recurring slug preview must change between occurrences.')
-    }
-
-    return errors
-  }, [creationMode, recurringOccurrencePreviews, scheduleOccurrenceDate])
-  const recurringEditorialWarnings = useMemo(() => {
-    if (creationMode !== 'recurring') {
-      return [] as string[]
-    }
-
-    const warnings = new Set<string>()
-    const [currentPreview, nextPreview] = recurringOccurrencePreviews
-
-    if (titleTemplate.trim() && !hasEventCreationDateTemplateVariable(titleTemplate)) {
-      warnings.add('Title template has no date variable, so recurring event titles may look identical between occurrences.')
-    }
-
-    if (form.resolutionRules.trim() && !hasEventCreationDateTemplateVariable(form.resolutionRules)) {
-      warnings.add('Resolution rules have no date variable, so recurring rules may look identical between occurrences.')
-    }
-
-    if (currentPreview && nextPreview && currentPreview.title.trim().toLowerCase() === nextPreview.title.trim().toLowerCase()) {
-      warnings.add('First and next recurring title previews are identical.')
-    }
-
-    if (currentPreview && nextPreview && currentPreview.resolutionRules.trim().toLowerCase() === nextPreview.resolutionRules.trim().toLowerCase()) {
-      warnings.add('First and next recurring resolution rules previews are identical.')
-    }
-
-    return Array.from(warnings)
-  }, [creationMode, form.resolutionRules, recurringOccurrencePreviews, titleTemplate])
-  const recurringRequiresServerWalletSetup = creationMode === 'recurring' && !hasConfiguredServerSigners
+  const {
+    scheduleDateValue,
+    scheduleOccurrenceDate,
+    recurrenceIntervalNumber,
+    hasRecurringDeployHistory,
+    automaticDeployAtIso,
+    automaticDeployAtDate,
+    nextRecurringResolutionDate,
+    nextRecurringDeployDate,
+    recurringResolvedTitle,
+    effectiveRecurringSlugTemplate,
+    recurringResolvedSlug,
+    recurringResolvedRules,
+    effectiveResolutionRules,
+    recurringOccurrencePreviews,
+    recurringPreviewErrors,
+    recurringEditorialWarnings,
+    recurringRequiresServerWalletSetup,
+  } = useRecurringEventPreview({
+    clientNowMs,
+    creationMode,
+    creatorSlugTail,
+    form,
+    hasConfiguredServerSigners,
+    initialDraftRecord,
+    recurrenceInterval,
+    recurrenceUnit,
+    slugSeed,
+    slugTemplate,
+    titleTemplate,
+  })
 
   const stepLabels = useMemo(
     () => ['Event', 'Market Structure', 'Resolution', 'Pre-sign', 'Sign & Create'],
     [],
   )
-  const previewEndDate = useMemo(() => {
-    const normalizedEndDate = normalizeDateTimeLocalValue(form.endDateIso)
-    if (!normalizedEndDate) {
-      return 'Resolution date not set'
-    }
-    const parsed = new Date(normalizedEndDate)
-    if (Number.isNaN(parsed.getTime())) {
-      return normalizedEndDate
-    }
-    return parsed.toLocaleString()
-  }, [form.endDateIso])
-  const previewTitle = useMemo(
-    () => creationMode === 'recurring'
-      ? (recurringResolvedTitle || titleTemplate.trim() || 'Untitled event')
-      : (form.title || 'Untitled event'),
-    [creationMode, form.title, recurringResolvedTitle, titleTemplate],
-  )
-  const previewSlug = useMemo(
-    () => creationMode === 'recurring'
-      ? (recurringResolvedSlug || effectiveRecurringSlugTemplate || 'event-slug')
-      : (form.slug || 'event-slug'),
-    [creationMode, effectiveRecurringSlugTemplate, form.slug, recurringResolvedSlug],
-  )
-  const previewMarkets = useMemo(() => {
-    if (form.marketMode === 'binary') {
-      return [
-        {
-          key: 'binary',
-          title: previewTitle.trim(),
-          question: (previewTitle || form.binaryQuestion).trim(),
-          shortName: '',
-          outcomeYes: form.binaryOutcomeYes.trim() || 'Yes',
-          outcomeNo: form.binaryOutcomeNo.trim() || 'No',
-          imageUrl: eventImagePreviewUrl,
-        },
-      ]
-    }
-
-    if (form.marketMode === 'multi_multiple' || form.marketMode === 'multi_unique') {
-      return form.options.map((option, index) => ({
-        key: option.id || `option-${index + 1}`,
-        title: option.title.trim(),
-        question: option.question.trim(),
-        shortName: option.shortName.trim(),
-        outcomeYes: option.outcomeYes.trim() || 'Yes',
-        outcomeNo: option.outcomeNo.trim() || 'No',
-        imageUrl: optionImagePreviewUrls[option.id] ?? null,
-      }))
-    }
-
-    return []
-  }, [
-    eventImagePreviewUrl,
-    form.binaryOutcomeNo,
-    form.binaryOutcomeYes,
-    form.binaryQuestion,
-    form.marketMode,
-    form.options,
-    optionImagePreviewUrls,
+  const {
+    previewEndDate,
     previewTitle,
-  ])
-  const tradePreviewMarket = useMemo(
-    () => previewMarkets[0] ?? null,
-    [previewMarkets],
-  )
-  const previewEventUrl = useMemo(
-    () => `${previewSiteOrigin}/event/${previewSlug}`,
-    [previewSiteOrigin, previewSlug],
-  )
-  const isMultiMarketPreview = form.marketMode === 'multi_multiple' || form.marketMode === 'multi_unique'
+    previewSlug,
+    previewMarkets,
+    tradePreviewMarket,
+    previewEventUrl,
+    isMultiMarketPreview,
+  } = useEventPreview({
+    creationMode,
+    effectiveRecurringSlugTemplate,
+    eventImagePreviewUrl,
+    form,
+    optionImagePreviewUrls,
+    previewSiteOrigin,
+    recurringResolvedSlug,
+    recurringResolvedTitle,
+    titleTemplate,
+  })
 
   const pendingAiIssues = useMemo(
     () => contentCheckIssues.filter(issue => !bypassedIssueKeys.includes(getAiIssueKey(issue))),
